% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTD}
\alias{CTD}
\title{Comparative toxicogenomics database list}
\format{A list with 1076 elements, with names corresponding to compounds and
  list items corresponding to gene symbols}
\description{
A list object containing drug::gene interactions based on the CTD and
filtered by human interactions that match CasRN from DrugBank.
}
\note{
Data sources downloaded April 21, 2016.
}
\keyword{datasets}

