% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksRepo.R
\name{ksRepo}
\alias{ksRepo}
\title{ksRepo}
\usage{
ksRepo(ccgenes, database, boot.method = "compound", nboot = 10000)
}
\arguments{
\item{ccgenes}{A character vector of prioritized genes derived from a
case/control dataset. Examples include differential gene expression,
differential methylation, and mutation frequency.}

\item{database}{A named list object representing a gene::compound interaction
database. List items should be named with compound names and their contents
should be interacting genes.}

\item{boot.method}{A string specifying whether the 'compound' gene lists or
the case/control, 'instance' dataset should be permuted for bootstrapping.
Default is to permute the compound gene lists.}

\item{nboot}{An integer specifying the number of bootstrap samples for p-value
calculation.}
}
\value{
A data frame with C rows and 5 columns, where C is the number of
 compounds in the database. compound gives name of the compound. n.genes
 gives the number of genes associated with each compound. ks gives the raw ks
 statistic, boot.p gives the bootstrapped p-value, and boot.fdr gives the
 fdr-corrected bootstrapped p-value.
}
\description{
\code{ksRepo} performs KS-based repositioning of user supplied data. KS values
are assigned p-values by bootstrapping the user supplied datasets.
}
\examples{
1+1
}

